글 쓴 사람만 글을 삭제할 수 있도록 기능 추가 글쓰기 하단 버턴 flex로 정렬하기 글쓰기 전송 실패 시 자동으로 글 내용을 임시등록하는 부분 삭제

+++ src/main/java/com/web/controller/BoardController.java
@@ -11,6 +11,8 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import java.security.Principal;
+
 @Controller
 @RequestMapping("/board")
 @RequiredArgsConstructor
@@ -26,16 +28,16 @@ public class BoardController {
     @ResponseBody
     @PostMapping("/register")
 //    public String registerPost(BoardDto boardDto, BindingResult bindingResult, Model model){
-    public String registerPost(@RequestBody BoardDto boardDto, Model model){
+    public String registerPost(@RequestBody BoardDto boardDto, Principal principal, Model model){
 
 //            if(bindingResult.hasErrors()){
 //            return "register";
 //        }
 
-        Long id;
+        boardDto.setWriter(principal.getName());
 
         try {
-            id  = boardService.register(boardDto);
+            boardService.register(boardDto);
         } catch (Exception e){
             model.addAttribute("errorMessage", "요청 등록 중 에러가 발생하였습니다.");
         }
@@ -54,9 +56,14 @@ public class BoardController {
     }
 
     @DeleteMapping("/remove/{id}")
-    public void remove(@PathVariable("id") Long id) {
+    public void remove(@PathVariable("id") Long id, Principal principal) {
 
-        boardService.remove(id);
+        BoardDto boardDto = boardService.readOne(id);
+        if(principal.getName().equals(boardDto.getWriter())){
+            boardService.remove(id);
+        }else{
+//            boardService.remove(id);
+        }
 
     }
